Heavly refactor Result representation of CNTs, outsource table as single component and use same template for wafer or raw evalaution result
